Revert "Merge remote-tracking branch 'origin/user-settings-page' into friends-page"
This reverts commit e74db322b5c1b10edb31218096df0c9b0a2cf4c3, reversing
changes made to 92dc5a3064b354706117158dbfd411cf7bd595cb.
