+++ client/src/App.js
@@ -10,6 +10,7 @@ import Login from "./pages/SignIn";
 import LoginRoute from "./routes/LoginRoute";
 import SignUpRoute from "./routes/SignUpRoute";
 import SignUp from "./pages/SignUp";
+import UserContext from "./context/UserContext";
 
 function App() {
   return (
@@ -25,10 +26,12 @@ function App() {
               path={'/signup'}
               component={SignUp}
             />
-            <ProtectedRoute
-              path={'/'}
-              component={UserDashboard}
-            />
+            <UserContext>
+              <ProtectedRoute
+                path={'/'}
+                component={UserDashboard}
+              />
+            </UserContext>
           </Switch>
         </BrowserRouter>
       </AuthProvider>
