+++ client/src/components/header/Navbar.js
@@ -8,7 +8,6 @@ import logo from "../../assets/logo.png";
 import MenuTabs from "./MenuTabs";
 import {Menu, MenuItem} from "@material-ui/core/"
 import { UserModel } from "../../context/UserContext";
-import { useAuthState } from "../../context/context";
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -49,7 +48,7 @@ export default function Navbar() {
   const classes = useStyles();
   const [anchorEl, setAnchorEl] = useState(null);
   const isMenuOpen = Boolean(anchorEl);
-  const currentUser = useAuthState();
+  const { imageUrl } = useContext(UserModel);
 
   // Handling user setting modal
   const [openSettingDialogue, setOpenSettingDialogue] = useState(false);
