+++ client/src/pages/Landing.js
@@ -2,7 +2,7 @@ import React from 'react'
 import Button from "@material-ui/core/Button";
 import Cookies from 'js-cookie';
 import {useAuthState, useAuthDispatch} from '../context/context';
-import {logout} from '../context/actions';
+import {logout} from '../actions';
 import { createBrowserHistory } from 'history';
 
 const history = createBrowserHistory();
